refactor(ecstore): move MultipartOperations into set_disk::ops::multipart (backlog#818)
P2 of the SetDisks God-Object split (tracking backlog#815, depends on P0
backlog#816 and follows P1 backlog#817).

Give the Multipart operation family its own module home: relocate
set_disk/multipart.rs to set_disk/ops/multipart.rs and move the
MultipartOperations storage-api contract impl (~1,157 lines) beside its
list/quorum helpers. The contract stays implemented for SetDisks so its
associated-type bounds are unchanged (ecstore_contract_compat_test still
covers it).

Method bodies are moved verbatim — a sorted content comparison of the old
mod.rs impl block plus the old multipart.rs against the new module is
byte-identical modulo the `use super::*;` → `use super::super::*;` header
swap. The cross-module helpers the impl relies on (get_upload_id_dir,
read_parts, rename_part, strip_internal_multipart_metadata) already carry
crate::set_disk visibility and remain reachable from the deeper module, so no
visibility widening was required.

Runtime behavior is unchanged.
